mart_api/inventory_api/pyproject.toml
[tool.poetry]
name = "stock"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.2"
sqlmodel = "^0.0.22"
psycopg2 = {extras = ["binary"], version = "^2.9.10"}
aiokafka = "^0.11.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
psycopg = {extras = ["binary"], version = "^3.2.3"}
protobuf = "^5.28.3"
python-jose = "^3.3.0"
pytest-cov = "^6.0.0"
pytest = "^8.3.3"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
